+++ packages/xrpc-server/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @atproto/xrpc-server
 
+## 0.11.10
+
+### Patch Changes
+
+- Updated dependencies [[`f8267c3`](https://github.com/bluesky-social/atproto/commit/f8267c3ada1645a2472c81c1ba95157466108b6e)]:
+  - @atproto/lex-client@0.3.0
+
 ## 0.11.9
 
 ### Patch Changes